Moved TestServer to OpenTelemetry.Tests and renamed TestHttpServer. Linked into Benchmarks, ZipkinTraceExporterTests, & OpenTelemetry.Adapter.Dependencies.Tests.
