+++ test/OpenTelemetry.Tests/Implementation/Metrics/Aggregators/CounterAggregatorTest.cs
