+++ test/OpenTelemetry.Tests/Implementation/Metrics/Aggregators/LastValueAggregatorTest.cs
