+++ test/OpenTelemetry.Tests/Implementation/Metrics/Aggregators/MinMaxSumCountAggregatorTest.cs
