+++ test/OpenTelemetry.Tests/Implementation/Metrics/CounterCleanUpTests.cs
