+++ test/OpenTelemetry.Tests/Implementation/Resources/ResourceTest.cs
