+++ test/OpenTelemetry.Tests/Implementation/Resources/ResourcesTests.cs
