+++ test/OpenTelemetry.Tests/Implementation/Trace/Propagation/B3FormatTest.cs
