+++ test/OpenTelemetry.Tests/Implementation/Trace/StatusTest.cs
