+++ test/OpenTelemetry.Tests/Implementation/Utils/CanonicalCodeExtensions.cs
